move all entries to a single file
