+++ desktop/main.js
@@ -9,7 +9,7 @@ const contextMenu = require('electron-context-menu');
 const {autoUpdater} = require('electron-updater');
 const log = require('electron-log');
 const ELECTRON_EVENTS = require('./ELECTRON_EVENTS');
-const checkForUpdates = require('../src/lib/checkForUpdates');
+const checkForUpdates = require('../src/libs/checkForUpdates');
 
 /**
  * Electron main process that handles wrapping the web application.
