+++ index.js
@@ -1,13 +1,11 @@
 /**
- * Note: This file is the entry point used by React Native for the native platforms.
- * Webpack uses different entry points for web and desktop, located at web/index.js and desktop/index.js, respectively.
- */
-/**
  * @format
  */
 
 import {AppRegistry} from 'react-native';
 import App from './src/App';
 import {name as appName} from './app.json';
+import additionalAppSetup from './src/setup';
 
 AppRegistry.registerComponent(appName, () => App);
+additionalAppSetup();
