+++ src/setup/index.desktop.js
@@ -1,0 +1,8 @@
+import {AppRegistry} from 'react-native';
+import {name as appName} from '../../app.json';
+
+export default function () {
+    AppRegistry.runApplication(appName, {
+        rootTag: document.getElementById('root'),
+    });
+}
