+++ src/setup/index.native.js
@@ -1,0 +1,4 @@
+// No additional setup required for native platforms
+export default function () {
+    return null;
+}
